matgo/matgo-crawler/requirements.txt
aiosignal==1.3.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
fake-useragent==2.0.3
beautifulsoup4==4.12.3
urllib3==2.2.3
torch==2.5.1
transformers==4.47.0
pillow==11.0.0